backend-i/session_9/pyproject.toml
[project]
name = "session-9"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "typer (>=0.15.2,<0.16.0)"
]

[tool.poetry]
packages = [{include = "session_9", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
